Move dataflow_to_sql_converter fixture to main fixture set
The dataflow_to_sql_converter fixture was scoped to the package
containing the plan_conversion tests, but it's useful anywhere
we need an initialized DataflowToSqlQueryPlanConverter built on
our most common simple semantic manifest.

Since we're splitting up the plan_conversion tests into plan conversion
and query rendering, it makes sense to hoist this out to where we
define our more broadly available fixtures.
